Move flux profile estimator
