+++ gammapy/estimators/points/__init__.py
@@ -1,3 +1,4 @@
 from .lightcurve import *
 from .sensitivity import *
 from .core import *
+from .profile import *
