+++ gammapy/estimators/points/tests/test_profile.py
